backend/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pyjwt
loguru